Muzzle Effect, Player Now Damages Zombies, Spatial UI
When the player shoots 'sparks' come out of the rifle muzzle accompied by flashes of light. Shooting at a zombie now causes it to lose health and eventually be destroyed. When the player gets damaged, splatters of blood will pop up on the screen and fade away after a few seconds